requirements.txt
absl-py==2.3.0
accelerate==1.7.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airportsdata==20250622
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
beautifulsoup4==4.13.4
blake3==1.0.5
blis==1.3.0
braceexpand==0.1.7
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
cmake==4.0.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
compressed-tensors==0.10.1
confection==0.1.5
contourpy==1.3.2
csvw==3.5.1
ctranslate2==4.6.0
cuda-bindings==12.9.0
cuda-python==12.9.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
cytoolz==1.0.1
datasets==3.6.0
decorator==5.2.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
Distance==0.1.3
distro==1.9.0
dlinfo==2.0.0
dnspython==2.7.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
faster-whisper==1.1.1
ffmpeg==1.4
fiddle==0.3.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.2
frozenlist==1.6.2
fsspec==2024.12.0
future==1.0.0
g2p-en==2.1.0
gevent==25.5.1
gguf==0.17.1
gguf-parser==0.1.1
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
inflect==7.5.0
iniconfig==2.1.0
interegular==0.3.3
intervaltree==3.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jiwer==3.1.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
kokoro==0.9.4
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lark==1.2.2
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.30.3
libcst==1.8.0
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
lm-format-enforcer==0.10.11
logging==0.4.9.6
loguru==0.7.3
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
misaki==0.9.4
mistral_common==1.6.2
more-itertools==10.7.0
moshi==0.2.9
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
nemo-toolkit==2.3.2
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvidia-ml-py==12.575.51
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.22.0
openai==1.91.0
openai-whisper==20240930
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
optuna==4.3.0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.3.0
parso==0.8.4
partial-json-parser==0.2.1.1.post6
pathspec==0.12.1
peft==0.15.2
phonemizer-fork==3.3.2
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
preshed==3.0.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.24.4
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==20.0.0
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pynvml==12.0.0
pyparsing==3.2.3
pyproject-metadata==0.7.1
pyreadline3==3.5.4
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-rtmidi==1.5.8
pytorch-lightning==2.5.1.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rfc3986==1.5.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
ruamel.yaml==0.18.13
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.7.0
scikit_build_core==0.8.1
scipy==1.15.3
segments==2.3.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==80.9.0
shellingham==1.5.4
shiboken6==6.9.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.13.1
soupsieve==2.7
sox==1.5.0
soxr==0.5.0.post1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphn==0.1.12
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
texterrors==0.5.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toolz==1.0.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.7.2
torchvision==0.20.1+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typeguard==4.4.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.161
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
vllm==0.9.1
wandb==0.20.1
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.111
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
whisper==1.1.10
win32_setctime==1.2.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.23.0
zope.event==5.1
zope.interface==7.2
